Core: Moved Stippling settings/property from core to base
