+++ modules/base/CMakeLists.txt
@@ -38,6 +38,8 @@ set(HEADER_FILES
     include/modules/base/datastructures/disjointsets.h
     include/modules/base/datastructures/imagereusecache.h
     include/modules/base/datastructures/kdtree.h
+    include/modules/base/datastructures/stipplingsettings.h
+    include/modules/base/datastructures/stipplingsettingsinterface.h
     include/modules/base/io/binarystlwriter.h
     include/modules/base/io/datvolumesequencereader.h
     include/modules/base/io/datvolumewriter.h
@@ -121,6 +123,7 @@ set(HEADER_FILES
     include/modules/base/properties/layerinformationproperty.h
     include/modules/base/properties/meshinformationproperty.h
     include/modules/base/properties/sequencetimerproperty.h
+    include/modules/base/properties/stipplingproperty.h
     include/modules/base/properties/volumeinformationproperty.h
 )
 ivw_group("Header Files" ${HEADER_FILES})
@@ -158,6 +161,8 @@ set(SOURCE_FILES
     src/basemodule.cpp
     src/datastructures/disjointsets.cpp
     src/datastructures/imagereusecache.cpp
+    src/datastructures/stipplingsettings.cpp
+    src/datastructures/stipplingsettingsinterface.cpp
     src/io/binarystlwriter.cpp
     src/io/datvolumesequencereader.cpp
     src/io/datvolumewriter.cpp
@@ -236,6 +241,7 @@ set(SOURCE_FILES
     src/properties/layerinformationproperty.cpp
     src/properties/meshinformationproperty.cpp
     src/properties/sequencetimerproperty.cpp
+    src/properties/stipplingproperty.cpp
     src/properties/volumeinformationproperty.cpp
 )
 ivw_group("Source Files" ${SOURCE_FILES})
