+++ modules/base/src/properties/stipplingproperty.cpp
