+++ modules/basegl/include/modules/basegl/datastructures/linesettings.h
@@ -30,7 +30,7 @@
 
 #include <modules/basegl/baseglmoduledefine.h>
 #include <modules/basegl/datastructures/linesettingsinterface.h>
-#include <inviwo/core/datastructures/stipplingsettings.h>
+#include <modules/base/datastructures/stipplingsettings.h>
 
 namespace inviwo {
