+++ modules/opengl/depends.cmake
@@ -1,5 +1,6 @@
 #--------------------------------------------------------------------
 # Dependencies for OpenGL module
 set(dependencies 
+	InviwoBaseModule
 )
 set(EnableByDefault ON)
