+++ modules/plotting/depends.cmake
@@ -2,7 +2,6 @@
 # Dependencies for current module
 # List modules on the format "Inviwo<ModuleName>Module"
 set(dependencies
-    #InviwoOpenGLModule
     InviwoDataFrameModule  
     InviwoFontRenderingModule
 )
